==> 1_nn_post_fetures.ipynb <==
 """
 Данный файл обрабатывает данные для постов пользователя с помощью нейросети. 
 Идея - закодировать посты в виде эмбедингов и использовать на них метод понижения пространства что бы использовать как фичи
 для обучения ML модели. Использую лингвистическю модель DistilBertModel, т.к. она легче Bert и Roberta,
 но не сильно уступает им в качестве. В качестве метода понижения пространства - PCA
 """

#забираю данные
import pandas as pd
posts_info = pd.read_csv('post.csv')


from transformers import DistilBertModel  # https://huggingface.co/docs/transformers/model_doc/distilbert#transformers.DistilBertModel


def get_model(model_name):
    """
    Функция возвращает токенайзер и предобученую модель по ключу 'distilbert'
    """
    assert model_name in ['distilbert']

    checkpoint_names = {'distilbert': 'distilbert-base-cased'  # https://huggingface.co/distilbert-base-cased}

    model_classes = {'distilbert': DistilBertModel}

    return AutoTokenizer.from_pretrained(checkpoint_names[model_name]), model_classes[model_name].from_pretrained(checkpoint_names[model_name])

tokenizer, model = get_model('distilbert')

### Напишем класс датасет для постов

from torch.utils.data import Dataset
from torch.utils.data import DataLoader
from transformers import DataCollatorWithPadding


class PostDataset(Dataset):
    """
    Класс обрабатывает полученый текст с помощью токнайзера 'distilbert'
    для использования в этой моделе    
    """
    def __init__(self, texts, tokenizer):
        super().__init__()

        self.texts = tokenizer.batch_encode_plus(
            texts,
            add_special_tokens=True,
            return_token_type_ids=False,
            return_tensors='pt',
            truncation=True,
            padding=True
        )
        self.tokenizer = tokenizer

    def __getitem__(self, idx):
        return {'input_ids': self.texts['input_ids'][idx], 'attention_mask': self.texts['attention_mask'][idx]}

    def __len__(self):
        return len(self.texts['input_ids'])
    
#создадим экземпляр класса для датасета    
dataset = PostDataset(posts_info['text'].values.tolist(), tokenizer)
#укажем способ сбора батчей с паддингом
data_collator = DataCollatorWithPadding(tokenizer=tokenizer)
#сформируем даталоадер для модели
loader = DataLoader(dataset, batch_size=32, collate_fn=data_collator, pin_memory=True, shuffle=False)

import torch
from tqdm import tqdm


@torch.inference_mode()
def get_embeddings_labels(model, loader):
    """
    Функция получает на вход нейросеть и лоадер с пердобработаными постами
    Проводите обработку лоадера в режиме применения 
    На выходе получаем эмбединги для каждого поста
    """
    model.eval()
    
    total_embeddings = []
    
    for batch in tqdm(loader):
        batch = {key: batch[key].to(device) for key in ['attention_mask', 'input_ids']}

        embeddings = model(**batch)['last_hidden_state'][:, 0, :]

        total_embeddings.append(embeddings.cpu())

    return torch.cat(total_embeddings, dim=0)

#проверим возморжность работы видеокарты
device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')

print(device)
print(torch.cuda.get_device_name())

model = model.to(device)

#формируем эмбединги
embeddings = get_embeddings_labels(model, loader).numpy()

embeddings

#на получены эмбединги искользуем PCA и скратим пространство до 2 измерений
from sklearn.decomposition import PCA

centered = embeddings - embeddings.mean()

pca = PCA(n_components=2,random_state = 74)

pca_decomp = pca.fit_transform(centered)

pd.DataFrame(pca_decomp)


def transform_post(post, n_pca = 20, drop = True, ohe = True,lemma =True):
    """
    Функция получает на вход таблицу пост post.
    Опчионально проводит OneHotEncoding типов постов,
    Лемматизацию текстов, Tf-Idf и накладывает на полученые вектора  PCA
    
    :n_pca: кол-во изерений в новом пространстве, если 0 не проводит pca b tf-idf
    :drop: = True/False дропать ли колонку текст
    :ohe: = True/False проводить ли  OneHotEncoding колонки topic
    :lemma: =True/False проводить ли лемматизацию
    
    
    """
    post_ = post.copy()
    #wnl = WordNetLemmatizer() #закоментил что бы не качать т.к. он не нужен в этом буке
    
    #лемматизация текста
    if lemma == True:
        def preprocessing(line, token=wnl):
            line = line.lower()
            line = re.sub(r"[{}]".format(string.punctuation), " ", line)
            line = line.replace('\n\n', ' ').replace('\n', ' ')
            line = ' '.join([token.lemmatize(x) for x in line.split(' ')])
            return line
    else:
        preprocessing = None

    # учу OHE_Post
    if ohe == True:
        one_hot = pd.get_dummies(post_['topic'], prefix='topic', drop_first=True)
        post_  = pd.concat((post_.drop(['topic'], axis=1), one_hot), axis=1)
        post_transformd = post_
    elif ohe == False:
        post_transformd = post_.drop(['topic'],axis =1 )
    
    # tf-idf+ pca
    if n_pca > 0 :
        # провожу tf-idf для Post    
        tf = TfidfVectorizer(stop_words='english',
                             preprocessor=preprocessing, 
                             min_df = 5) #создаю экземпляр класса
        tf_idf_ = tf.fit_transform(post_['text'])#учу класс
        tf_idf_ = tf_idf_.toarray() - tf_idf_.mean() #центрируем данные

        list_col_pca = [f"PCA_{nn}" for nn in range(1,n_pca + 1)] 
        pca = PCA(n_components=n_pca,random_state = 74)
        
        # создаю экземплря PCA
        PCA_dataset = pca.fit_transform(tf_idf_) #провожу PCA 
        PCA_dataset = pd.DataFrame(PCA_dataset, columns=list_col_pca,index=post.index)
        
        # Трансформирую Post    
        post_transformd = pd.concat((post_transformd, PCA_dataset), axis=1)
        if drop == True:
            post_transformd = post_transformd.drop(['text'],axis=1) 
            
    
    else:
        if drop == True:
            post_transformd = post_transformd.drop(['text'],axis=1)
        
    return post_transformd

# таблица post без текса и раскодироваными topic
new_post_df = transform_post(posts_info,n_pca = 0, drop = True, ohe = True,lemma =False)

# обработаная таблица для обучения модели
new_post_df = pd.merge(new_post_df, pd.DataFrame(pca_decomp,columns = ['PCA_1', 'PCA_2']),left_index=True, right_index=True,how='left')

#сохраняем данные
new_post_df.to_csv('post_transform_nn.csv', index=False)
