==> 3_nn_save_data.ipynb <==
"""
Подготовавливаем данные из таблицы POST для залива в БД и дальнейшей работы модели
Текст заранее предобработан нейросетью в шаге 1_nn_post_fetures
"""

import pandas as pd
pd.set_option('display.max_columns', 500)
import numpy as np
import url from url


from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker


post = pd.read_csv('post_pca_2_nn.csv')



post_original = pd.read_csv('post.csv')

post.insert(loc = 1,
          column = 'text',
          value = post_original['text'])

post.insert(loc = 2,
          column = 'topic',
          value = post_original['topic'])

post

engine = create_engine(url)

#записал эту таблицу
post.to_sql('shestov_post_lesson_22_2_pca_nnn', con=engine) # записываем таблицу
